+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Store food filters in user profile [#1101](https://github.com/rokwire/illinois-app/issues/1101).
 - Import stored user profile and settings on first app launch [#1103](https://github.com/rokwire/illinois-app/issues/1103).
 - Make "Kognito At Risk (Counseling Center)" wellness buttons to launch "kognito" guide detail page [#1105](https://github.com/rokwire/illinois-app/issues/1105).
+- Load guide detail panels on "Counseling" and "ACE IT" buttons from Mental Wellness panel [#1107](https://github.com/rokwire/illinois-app/issues/1107).
 
 ## [3.0.63] - 2021-12-08
 ### Added